css/css-scroll-snap/parsing/scroll-padding-bottom-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding-bottom with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding-bottom">
<meta name="assert" content="scroll-padding-bottom supports the full grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-padding-bottom", "auto");

test_valid_value("scroll-padding-bottom", "10px");
test_valid_value("scroll-padding-bottom", "20%");
test_valid_value("scroll-padding-bottom", "calc(2em + 3ex)");
test_valid_value("scroll-padding-bottom", "calc(50% + 60px)");
</script>
</body>
</html>
